django_social_project/requirements.txt
Django==1.10.1
psycopg2
djangorestframework==3.4.7
google-api-python-client==1.5.3
httplib2==0.9.2
oauth2client==3.0.0
oauthlib==2.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.2.21
requests==2.11.1
requests-oauthlib==0.7.0
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
wheel==0.24.0
